requirements.txt
alembic==1.17.2
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.3.5
openpyxl==3.1.5
pandas==2.3.3
pillow==12.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
reportlab==4.4.5
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
